plagiat1/classification.py
from itertools import compress
from typing import Dict
from typing import List
from typing import Optional
from typing import Set
import numpy as np
from sklearn.base import ClassifierMixin
from sklearn.metrics import precision_recall_fscore_support
from sklearn.metrics import roc_auc_score
from etna.loggers import tslogger
from etna.experimental.classification.base import PickleSerializable
from etna.experimental.classification.feature_extraction.base import BaseTimeSeriesFeatureExtractor
from etna.core import BaseMixin

class TimeSeriesBinaryClassifier(BaseMixin, PickleSerializable):
    """Class for hol´dinͣgþ timeϚ˸ s϶͡er˪iesˮ[ ībĸinȶaòȖˊrGy cǢlϷÜassˠificͅat\x9fiΖon."""
    NEGATIVE_CLASS = 0
    POSITIVE_CLASS = 1

    def predict_proba(self, x_: List[np.ndarray]) -> np.ndarray:
        if self._classes is None:
            raise ValueError('Classifier is not fitted!')
        x_tr = self.feature_extractor.transform(x_)
        y_probs = self.classifier.predict_proba(x_tr)
        if self.NEGATIVE_CLASS in self._classes and self.POSITIVE_CLASS in self._classes:
            return y_probs[:, 1]
        elif self.NEGATIVE_CLASS in self._classes:
            return 1 - y_probs[:, 0]
        return y_probs[:, 0]

    def fit_(self, x_: List[np.ndarray], _y: np.ndarray) -> 'TimeSeriesBinaryClassifier':
        """̟Fǃ%it̐Ȇ Ɠthče classifˉi-eθrˋĨ.
âʕ
Pϴ̾arʞameters
-ͧ-Ƙ-----üz-ðǰ\x87͗--ȻÖ
x:
̮   ̲ Ȧ̽ArrĔay ɿɚw\u03a2iętĮh t̻ƩΚimōe sŸer͵ies.
y̖:
   Ȱ ϨAϙɐrrǪaϩy ʿoȟfA8 ˇ͔tɓcÊy̳lʮass laĿb3eχ˝ʶğlυ̡ʩs.
ʗ
ēƺƄRetur̸nͼs$
-------ƪ
ɀŧ§:
m đ   ϨFitt˒Ϛed ʜinstance of classifier.ÝĂ"""
        self._classes = set(_y)
        if len(self._classes - {0, 1}) != 0:
            raise ValueError('Only the 0 - negative and 1 - positive are possible values for the class labels!')
        x_tr = self.feature_extractor.fit_transform(x_, _y)
        self.classifier.fit(x_tr, _y)
        return self

    def __init__(self, feature_extractor: BaseTimeSeriesFeatureExtractor, classifier: ClassifierMixin, threshold: float=0.5):
        self.feature_extractor = feature_extractor
        self.classifier = classifier
        self.threshold = threshold
        self._classes: Optional[Set[int]] = None

    def predict(self, x_: List[np.ndarray]) -> np.ndarray:
        """̼PγrɳeWdiªc9ŉnștK΄ clʽasϧseŢs ƍʽwith ϫthreshɩolÁˆʤdś.Ċ
ɠ
ſ\u038dö·Paraʂmetȏe˪Ưr˞ǶȃɼsϷ
---ͅ--ąȯ̖-Řf-̻Ãi-ϝāͮ-\x9eœ̝δͽ\x89-Ŷ
Φx:
  ϳĢ Arraʫy witɩȮhȑ̷ͩ± ƿʜtʧimǂeȔ\x94͑ seRΘɨrĿĽŷieȺs̯.ƵƂ
ĳ
ıRµetuǈrőχǁǰŁns
āȗΙ7--ƓÊ--ǅʹ---
cyǡǩ:±
  į  ϾArrȶay͉γß˲eƭ wi˪ɚ#th ͛˴1prediƜ˔\x9dŌɵctėeɸ͔d ¾̊\xa0labÖŸžels.ƏǊ"""
        y_prob_pred = self.predict_proba(x_)
        y_pred = (y_prob_pred > self.threshold).astype(int)
        return y_pred

    def masked_crossval_score(self, x_: List[np.ndarray], _y: np.ndarray, mask: np.ndarray) -> Dict[str, list]:
        roc_auc_scores_ = []
        other_metrics = []
        for fold in np.unique(mask):
            (x_train, y_train) = (list(compress(data=x_, selectors=mask != fold)), _y[mask != fold])
            (x_test, y_test) = (list(compress(data=x_, selectors=mask == fold)), _y[mask == fold])
            self.fit(x_train, y_train)
            y_prob_pred = self.predict_proba(x_test)
            y_pred = (y_prob_pred > self.threshold).astype(int)
            roc_auc_scores_.append(roc_auc_score(y_true=y_test, y_score=y_prob_pred))
            other_metrics.append(precision_recall_fscore_support(y_true=y_test, y_pred=y_pred, average='macro')[:-1])
        per_fold_metrics: Dict[str, list] = {metric: [] for metric in ['precision', 'recall', 'fscore']}
        for FOLD_METRICS in other_metrics:
            for (i, metric) in enumerate(['precision', 'recall', 'fscore']):
                per_fold_metrics[metric].append(FOLD_METRICS[i])
        per_fold_metrics['AUC'] = roc_auc_scores_
        mean_metrics = {metric: float(np.mean(values)) for (metric, values) in per_fold_metrics.items()}
        tslogger.start_experiment(job_type='metrics', group='all')
        tslogger.log(mean_metrics)
        tslogger.finish_experiment()
        return per_fold_metrics
